rename files based on convention
