+++ backend/src/bounded-contexts/iam/iam/repository/user-write.pg.repository.ts
@@ -8,7 +8,7 @@ import {
 } from '@bitloops/bl-boilerplate-core';
 import { Injectable, Inject } from '@nestjs/common';
 import * as jwtwebtoken from 'jsonwebtoken';
-import { UserWriteRepoPort } from '@src/lib/bounded-contexts/iam/authentication/ports/UserWriteRepoPort';
+import { UserWriteRepoPort } from '@src/lib/bounded-contexts/iam/authentication/ports/user-write.repo-port';
 import { UserEntity } from '@src/lib/bounded-contexts/iam/authentication/domain/UserEntity';
 import { constants } from '@bitloops/bl-boilerplate-infra-postgres';
 import { EmailVO } from '@src/lib/bounded-contexts/iam/authentication/domain/EmailVO';
