+++ backend/src/bounded-contexts/marketing/marketing/service/mock-email.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@nestjs/common';
 import {
   EmailServicePort,
   SendEmailRequest,
-} from 'src/lib/bounded-contexts/marketing/marketing/ports/email-service-port';
+} from '@src/lib/bounded-contexts/marketing/marketing/ports/email.service-port';
 
 @Injectable()
 export class MockEmailService implements EmailServicePort {
