+++ backend/src/bounded-contexts/todo/todo/todo.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { TodoWriteRepository } from './repository/todo-write.repository';
 import { TodoReadRepository } from './repository/todo-read.repository';
 import { TodoModule as LibTodoModule } from 'src/lib/bounded-contexts/todo/todo/todo.module';
-import { TodoWriteRepoPortToken } from '@src/lib/bounded-contexts/todo/todo/ports/TodoWriteRepoPort';
-import { TodoReadRepoPortToken } from '@src/lib/bounded-contexts/todo/todo/ports/TodoReadRepoPort';
+import { TodoWriteRepoPortToken } from '@src/lib/bounded-contexts/todo/todo/ports/todo-write.repo-port';
+import { TodoReadRepoPortToken } from '@src/lib/bounded-contexts/todo/todo/ports/todo-read.repo-port';
 import { MongoModule } from '@bitloops/bl-boilerplate-infra-mongo';
 import { PubSubCommandHandlers } from '@src/lib/bounded-contexts/todo/todo/application/command-handlers';
 import { PubSubQueryHandlers } from '@src/lib/bounded-contexts/todo/todo/application/query-handlers';
