+++ backend/src/lib/bounded-contexts/iam/authentication/application/command-handlers/log-in.handler.ts
@@ -10,7 +10,7 @@ import { LogInCommand } from '../../commands/log-in.command';
 import {
   UserWriteRepoPortToken,
   UserWriteRepoPort,
-} from '../../ports/UserWriteRepoPort';
+} from '../../ports/user-write.repo-port';
 import { ApplicationErrors } from '../errors';
 import { Traceable } from '@bitloops/bl-boilerplate-infra-telemetry';
