+++ backend/src/lib/bounded-contexts/marketing/marketing/ports/email.service-port.ts
