+++ backend/src/lib/bounded-contexts/marketing/marketing/tests/__tests__/todo-completions-incremented/notification-template-read-repo.mock.ts
@@ -6,7 +6,7 @@ import {
   asyncLocalStorage,
 } from '@bitloops/bl-boilerplate-core';
 import { NotificationTemplateReadModel } from '../../../domain/read-models/notification-template.read-model';
-import { NotificationTemplateReadRepoPort } from '../../../ports/notification-template-read.repo-port.';
+import { NotificationTemplateReadRepoPort } from '../../../ports/notification-template-read.repo-port';
 import {
   SUCCESS_CASE,
   UNSUCCESS_REPO_ERROR_CASE,
