+++ backend/src/lib/bounded-contexts/todo/todo/application/command-handlers/delete-todo.handler.ts
@@ -11,7 +11,7 @@ import { ApplicationErrors } from '../errors';
 import {
   TodoWriteRepoPort,
   TodoWriteRepoPortToken,
-} from '../../ports/TodoWriteRepoPort';
+} from '../../ports/todo-write.repo-port';
 import { Traceable } from '@bitloops/bl-boilerplate-infra-telemetry';
 
 type DeleteTodoUseCaseResponse = Either<
