+++ backend/src/lib/bounded-contexts/todo/todo/tests/__tests__/get-todos/get-todos-read-repo.mock.ts
@@ -6,7 +6,7 @@ import {
   asyncLocalStorage,
 } from '@bitloops/bl-boilerplate-core';
 import { TodoReadModel } from '../../../domain/TodoReadModel';
-import { TodoReadRepoPort } from '../../../ports/TodoReadRepoPort';
+import { TodoReadRepoPort } from '../../../ports/todo-read.repo-port';
 import {
   GET_TODOS_REPO_ERROR_CASE,
   GET_TODOS_SUCCESS_CASE,
