+++ backend/src/lib/bounded-contexts/todo/todo/tests/__tests__/modify-title-todo/modify-title-write-repo.mock.ts
@@ -6,7 +6,7 @@ import {
   fail,
 } from '@bitloops/bl-boilerplate-core';
 import { TodoEntity } from '@src/lib/bounded-contexts/todo/todo/domain/TodoEntity';
-import { TodoWriteRepoPort } from '@src/lib/bounded-contexts/todo/todo/ports/TodoWriteRepoPort';
+import { TodoWriteRepoPort } from '@src/lib/bounded-contexts/todo/todo/ports/todo-write.repo-port';
 import {
   MODIFY_INVALID_TITLE_CASE,
   MODIFY_TITLE_SUCCESS_CASE,
